feat: create route for match view
