+++ src/components/ActionBtn/index.jsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import './actionbtn.scss'
-const UnirseBtn = ({action}) => {
+const index = ({action}) => {
   return (
-    <div className="text-center sticky-md-top">
+    <div className="text-center w-100">
       <button type="button" className="action-btn btn ml-auto mr-auto" >{action}</button>
     </div>
   )
 }
 
-export default UnirseBtn
+export default index
