+++ src/components/JugadorEquipo/index.jsx
@@ -1,8 +1,10 @@
-import avatar from 'assets/user_avatar.png'
+import avatar from 'assets/images/user_avatar2.jpg'
+import './jugadorequipo.scss'
 const JugadorEquipo = (team) => {
   return (
     <div className="d-flex flex-column justify-content-center align-items-center jugador">
-      <img src={avatar} alt="player1" />
+      {/* <img className="user-img" src={avatar} alt="player1" /> */}
+      <div className="user-img" style={{backgroundImage: `url(${avatar})`}}></div>
       <h5 className="mt-1 mb-0">Jugador 1</h5>
       <p>Posicion</p>
