+++ src/pages/VistaPartido/vistapartido.scss
